Settings sliders change value immediately (closes #2)
Previously the "onchange" event was used, which fires
when the slider was changed and the mouse button released.
Now we additionally look for the "oninput" event, which
fires each time the value is changed (before releasing
the mouse). The onchange event is still observed because
IE10 doesn't support oninput. (tnx suggestion WA2NFN)